Nonworking approach to restructure renderers
